#  ENH: making mode stable/keeping original ordering