@jreback 2021-01-25
can you run the mode asvs (doubt we have too many), value_counts & some indexing ones just to make sure nothing is changing here