@realead 2021-01-25
I have added some asv-benchmarks, because otherwise mode seems to be not covered in asv_bench.

For running `asv continuous -f 1.10 upstream/master HEAD  -b series_methods.Mode -b ^indexing` I got:

```
       before           after         ratio
     [309cf3a6]       [e91e27a2]
+        18.6±1ms         51.3±4ms     2.76  series_methods.Mode.time_mode(100000, 'int')
+      18.9±0.5ms         38.4±5ms     2.03  series_methods.Mode.time_mode(100000, 'uint')
+      43.8±0.4ms         88.9±7ms     2.03  series_methods.Mode.time_mode(100000, 'float')
+     1.19±0.03ms      1.55±0.02ms     1.30  series_methods.Mode.time_mode(10000, 'int')
+     1.20±0.01ms      1.52±0.03ms     1.26  series_methods.Mode.time_mode(10000, 'uint')
+     1.89±0.08ms      2.39±0.04ms     1.26  series_methods.Mode.time_mode(10000, 'float')
+         134±2ms          153±2ms     1.15  series_methods.Mode.time_mode(100000, 'object')
```

There is a slow down, as soon as the table with keys no longer fits the cache (10^3 is OK, 10^5 isn't). More cache-misses should be expected, due to an additional indirection in `value_counts_xxx`, which is needed for keeping the original order.

Looking at the results it looks as if the array-sizes in asv-tests for `value_counts` are too small, otherwise this regression should be visible in asv-run in #39009 as well.