@realead 2021-01-25
When I change `series_methods.ValueCounts` similar to https://github.com/pandas-dev/pandas/pull/39353/commits/ecf9ff6e0702f54e41537cdc58a4e4675b6b23d3, i.e. adding different array-sizes, and compare with asv PR #39009 I get the following:

```
       before           after         ratio
     [2f883215]       [ecf9ff6e]
+      25.6±0.8ms         74.5±1ms     2.91  series_methods.ValueCounts.time_value_counts(100000, 'int')
+        26.4±1ms         73.1±5ms     2.76  series_methods.ValueCounts.time_value_counts(100000, 'uint')
+      51.2±0.7ms         123±20ms     2.40  series_methods.ValueCounts.time_value_counts(100000, 'float')
-         212±9ms          176±3ms     0.83  series_methods.ValueCounts.time_value_counts(100000, 'object')
-     1.50±0.02ms      1.05±0.02ms     0.70  series_methods.ValueCounts.time_value_counts(1000, 'object')
-      12.5±0.1ms       8.57±0.1ms     0.69  series_methods.ValueCounts.time_value_counts(10000, 'object')
```

With smaller sizes (as it is currently the case), the issues with cache misses aren't visible and one sees only  speed-up in case of `object` (because some pre-/post-processing is no longer needed):

```
       before           after         ratio
     [2f883215]       [ecf9ff6e]
-     1.50±0.02ms      1.05±0.02ms     0.70  series_methods.ValueCounts.time_value_counts(1000, 'object')
```