@jreback 2021-01-26
is the build_count_table just a special case version of value_counts that doesn't keep order (and that is why its faster)?