@realead 2021-01-26
@jreback 

> is the build_count_table just a special case version of value_counts that doesn't keep order (and that is why its faster)?

that is correct. However, my implementation is  a little bit naive: I have missed, that there are more cache misses now (misses for look-up in table + miss for lookup in the line `result_counts.data.data[unique_key_index] += 1` vs misses only for look-up table in the old version) and it really counts for bigger arrays.

I think there is a way to improve the performance (which would also improve `value_counts` as well) making it almost as fast as the old version. I will try it.