@realead 2021-01-26
Comparison to the state before #39009 was merged (`-b series_methods.ValueCounts  -b series_methods.Mode`):

```
       before           after         ratio
     [2f883215]       [31512ea5]
-         251±5ms          197±4ms     0.79  series_methods.ValueCounts.time_value_counts(100000, 'object')
-      13.3±0.8ms       9.36±0.8ms     0.70  series_methods.ValueCounts.time_value_counts(10000, 'object')
-      1.73±0.1ms      1.18±0.08ms     0.68  series_methods.ValueCounts.time_value_counts(1000, 'object')
```

no additional cache misses and better times for object, because post processing was removed.